+++ src/metrics/tokens.rs
@@ -169,6 +169,7 @@ implement_metric_trait!(
     TypescriptCode,
     TsxCode,
     CppCode,
+    MozcppCode,
     RustCode,
     PreprocCode,
     CcommentCode,